Create longestAphabeticalSubstring0.py
A first attempt to solve the problem of finding the longest alphabetical substring.

Utilizes two pointers: 'start' and 'end'. The 'end' pointer advances until it encounters a letter that comes before the previous one. 

When this occurs, the algorithm determines if the current substring is longer than any previous ones.